+++ hub.js
@@ -1,15 +1,21 @@
 "use strict";
 
+// Imports
+
 const fs = require('fs');
 const os = require('os');
 const net = require('net');
 const http = require('http');
 const express = require('express');
-const EventEmitter = require('events');
+const emitter = require('events');
 const env = require('dotenv').config();
+const httpProxy = require('http-proxy');
 const cookies = require('cookie-parser');
+const exec = require('child_process').exec;
 const sessions = require('express-session');
 const sessionFile = require('session-file-store')(sessions);
+
+// Create sessions
 
 const session = sessions({
   secret: process.env.COOKIE_SECRET,
@@ -18,386 +24,338 @@ const session = sessions({
   store: new sessionFile()
 });
 
-let server = express();
+// Create global for accessing sessions
+let userSession;
 
+// Create app server
+
+let server = express();
 server.use(session);
 server.use(cookies());
 
-let sess;
-let interrupt;
-
-// Listen for incoming traffic on port 8080
+// Listen for incoming traffic
 
 const app = http.createServer(server);
 app.listen(8080);
 
-// Create registries (ports occupied, containers running)
+// Create reserved port registry
+let ports = [8000, 8080];
 
-let ports = [80, 443, 4180, 5000, 8000, 8080];
-let registry = { };
+// Docker daemon setup
 
-// Docker setup
+const docker = require("dockerode");
+const moby = new docker({socketPath: '/var/run/docker.sock'});
 
-const Docker = require("dockerode");
-const ishmael = new Docker({socketPath: '/var/run/docker.sock'});
+// Setup event emitter
 
-// Create event emitter
+const events = new emitter();
 
-const emitter = new EventEmitter();
+// Setup global activity tracker
 
-/**
- * Creates random port assignment between 1000 and 65535
- * @function randomize
- * @private
- * @param {String}  lower   Lower limit of random range
- * @param {String}  upper   Upper limit of random range
- */
-const randomize = (lower, upper) => {
-  return Math.floor(Math.random() * (upper - lower) + lower);
-}
-
-const now = () => {
-  return Math.floor(new Date().getTime() / 1000);
-}
+let activity = {}
 
 /**
- * Discovers ports already in use
- * @function occupied
- * @private
- * @param {String}    port  Port to query
- * @param {Function}  fn    Callback
+ * Event used to store last active times
+ * @param {Object} store    Packet of active state info
  */
-const occupied = (port) => {
-  let server = net.createServer((socket) => {
+events.on('updateLastActive', (store) => {
+  activity[store.user] = { "lastActive": store.lastActive };
+});
+
+/**
+ * Discover ports already in use
+ * @function isOccupiedPort
+ * @private
+ * @param {String}  port  Port to query
+ */
+const isOccupiedPort = (port) => {
+  let local = net.createServer( (socket) => {
     socket.write('Ping\r\n');
     socket.pipe(socket);
   });
-  server.on("error", (err) => {
+  local.on("error", (err) => {
     return true;
   });
-  server.on("listening", (success) => {
-    server.close();
+  local.on("listening", (success) => {
+    local.close();
     return false;
   });
-  server.listen(port, '0.0.0.0');
-}
+  local.listen(port, "0.0.0.0");
+};
 
 /**
- * Generates a unique port for new containers
- * @function port
+ * Creates random port assignment 1000-65535
+ * @function randomPort
  * @private
+ * @param {String} lower  Lower limit of random range
+ * @param {String} upper  Upper limit of random range
  */
-const port = () => {
-  let pid = randomize(1000, 65535);
-  while(true) {
-    let used = occupied(port);
-    if(!ports.hasOwnProperty(pid)) {
-      ports.push(pid);
-      break;
-    }
-    pid = randomize(1000, 65535);
-  }
-  return pid;
-}
+const randomPort = (lower, upper) => {
+  let port;
+  do {
+    port = Math.floor(Math.random() * (upper - lower) + lower);
+  } while(isOccupiedPort(port));
+  ports.push(port);
+  return port;
+};
 
 /**
- * Acquires address from a container's properties
- * @function address
+ * Get current time on request
+ * @function now
  * @private
- * @param {Container} container Instance of an individual container
- * @param {Function}  fn        Callback function
  */
-const address = (container, fn) => {
-  container.inspect((err,data) => {
-    let addr = data.NetworkSettings.Networks.bridge.IPAddress;
-    if(!addr) { address(container,fn) }
-    else { fn(addr) }
-  });
-}
+const now = () => {
+  return Math.floor(new Date().getTime() / 1000);
+};
 
+/**
+ * Retrieve container data from running container(s)
+ * @function containerData
+ * @param {String} user   Authenticated user requesting container
+ */
 const containerData = async (user) => {
   let container;
-  let containers = await ishmael.listContainers({names: user});
-  for await(let entry of containers){
-    let acquired = await ishmael.getContainer(entry.Id);
+  let containers = await moby.listContainers({names: user});
+  for await(let entry of containers) {
+    let acquired = await moby.getContainer(entry.Id);
     container = await acquired.inspect();
+    return await {
+      id: container.Id,
+      port: container.NetworkSettings.Ports['8000/tcp'][0].HostPort
+    }
   }
   return await {
-    port: container.NetworkSettings.Ports['8000/tcp'][0].HostPort,
-  };
+    id: null,
+    port: null
+  }
+};
+
+/**
+ * Read global user directory for user details
+ * @function readDirectory
+ * @private
+ */
+let directory;
+
+const readDirectory = () => {
+  let json = fs.readFileSync(process.env.DIRECTORY);
+  directory = JSON.parse(json);
 }
 
 /**
- * Attempts to connect to the container on the generated port
- * @function connect
+ * Start a new container for users requiring one
+ * @function startContainer
  * @private
- * @param {String}    user  Username of user from x-forwarded-user
- * @param {Function}  fn    Callback function
+ * @param {String} user       User requesting a container
+ * @param {String} port       Port binding for new container
+ * @param {Object} directory  Directory of relevant user data
  */
-const connect = async (user, fn) => {
-  let world = await containerData(user);
-  http.get({ host: "0.0.0.0", port: world.port, path: `/` }, (res) => {
-    fn();
-  }).on('error', (err) => {
-    connect(user, fn);
-  });
-};
-
-/**
- * Queries docker container pool for users with running containers
- * @function alive
- * @private
- * @param {String}    user  Username of user from x-forwarded-user
- */
-const alive = async (user) => {
-  let list = await ishmael.listContainers({all: true});
-  for(let entry of list) {
-    let names = entry.Names;
-    if(names.includes(`/${user}`)){
-      return true
-    }
-  }
-  return false;
-};
-
-// Event used to add information to global container registry
-
-emitter.on('register', (store) => {
-  if(!registry[store.user]) registry[store.user] = { "params": { } };
-  for(let param in store.params) {
-    registry[store.user].params[param] = store.params[param]
-  }
-});
-
-// Read directory to pass the DISTRICT environment variable
-let directory;
-
-fs.readFile(process.env.DIRECTORY, (err, data) => {
-  directory = JSON.parse(data);
-});
-
-// Set proxy object for web and socket proxies
-
-const httpProxy = require('http-proxy');
-
-/**
- * Acquires content at /login endpoint
- * @param {Object}  req   Web request
- * @param {Object}  res   Web response
- */
-server.get('/login', async(req, res) => {
-  // Acquire registry
-  fs.readFile(process.env.DIRECTORY, (err, data) => {
-    directory = JSON.parse(data);
-  });
-  // Acquire random port
-  let pid = port();
-  // Get authenticated user from header or session
-  let user;
-  session(req, {}, () => {
-    user =  req.headers['x-forwarded-user'] || req.session.user;
-    // Set user property of session
-    sess = req.session;
-    sess.user = user;
-  });
-  // If user is already attached to a container
-  let isAlive = await alive(user);
-  if(isAlive){ connect(user, () => { res.redirect("/"); res.end(); });  }
-  // Create container from Docker API
-  let userId = directory[user].uid;
+const startContainer = (user) => {
+  let uid = directory[user].uid;
+  let gid = directory[user].gid
   let district = directory[user].district;
-  let districtId = directory[user].gid;
-  ishmael.run(`world:${process.env.IMAGE}`, [], undefined, {
+  moby.run(`world:${process.env.IMAGE}`, [], undefined, {
     "name": `${user}`,
     "Hostname": "term-world",
     "Env": [
       `VS_USER=${user}`,
-      `VS_USER_ID=${userId}`,
+      `VS_USER_ID=${uid}`,
       `DISTRICT=${district}`,
-      `GID=${districtId}`
+      `GID=${gid}`
     ],
-    "ExposedPorts": {"8000/tcp":{}},
+    "ExposedPorts": { "8000/tcp":{} },
     "HostConfig": {
       "Binds": [`${process.env.VOLUME}:/world`],
       "PortBindings": {
         "8000/tcp": [
           {
-            "HostPort": pid.toString()
+            "HostPort": `${randomPort(1000,65535)}`
           }
         ]
       }
     }
-  }, (err,data,container) => {
-    if(err) { console.log(err); }
-  }).on('container', (container) => {
-    // On container creation, get container private address
-    address(container, (addr) => {
-      emitter.emit('register',
-        {
-          user: user,
-          params: {
-            port: pid,
-            sockets: 0,
-            address: addr,
-            container: container
-          }
-        }
-      );
-      // Callback to redirect request
-      connect(user, () => {
-        res.redirect(`/`);
-      });
-    })
+  }, (err, data, container) => {
+    if(err) throw err;
   });
-});
+}
 
 /**
- * Acquires content at / endpoint
- * @param {Object}  req   Web request
- * @param {Object}  res   Web response
+ * Attempt connection to requested container on generated port
+ * @function connectContainer
+ * @private
+ * @param {String} user   Authenticated user requresting container
  */
-server.get('/*', (req,res) => {
-  let user;
-  session(req, {}, () =>  {
-    user = req.session.user;
+const connectContainer = async (user, callback) => {
+  let world = await containerData(user);
+  http.get({
+    host: "0.0.0.0",
+    port: world.port,
+    path: "/"
+  }, (res) => {
+    callback();
+  }).on("error", (err) => {
+    connectContainer(user, callback);
   });
-  if(
-    user === undefined ||
-    registry[user] === undefined
-  ) {
-    return res.redirect('/login');
+};
+
+/**
+ * Answers /login endpoint, creates containers for new users; rejoins old
+ * @param {Object} req    Web request
+ * @param {Object} res    Web response
+ */
+server.get("/login", async (req, res) => {
+  // Grab or create session data
+  let user;
+  session(req, {}, () => {
+    user = req.headers["x-forwarded-user"] || req.session.user;
+    userSession = req.session;
+    userSession.user = user;
+  });
+  // If user has a container running, boot to that container
+  let world = await containerData(user);
+  // Otherwise, create a new container and proceed
+  if(world.id === null) {
+    // Get details necessary to start container
+    readDirectory();
+    startContainer(
+      user
+    );
   }
-  const proxy = httpProxy.createServer({});
-  proxy.web(req, res,
-    {target: `http://localhost:${registry[user].params.port}/`}
-  );
-  proxy.on("error", (err, req, res) => {
-    console.log(`PROXY ERROR (1): ${err}`);
-    delete registry[user];
-    res.redirect("/login");
+  await connectContainer(user, () => {
+    res.redirect("/")
   });
 });
 
 /**
- * Handles transfer of HTTP protocol to web sockets
- * @param {Object}  req     Web request
- * @param {Object}  socket  Socket created
- * @param {Object}  head    ?
+ * Answers / endpoint for created containers
+ * @param {Object} req    Web request
+ * @param {Object} res    Web response
  */
-
-app.on('upgrade', (req, socket, head) => {
+server.get("/*", async (req, res) => {
   let user;
-  let proxy = httpProxy.createServer({});
   session(req, {}, () => {
     user = req.session.user;
-    if(user === undefined || registry[user] === undefined) { return; }
-    proxy.ws(req, socket, head,
-      {target: `http://localhost:${registry[user].params.port}/`}
-    );
-    registry[user].params.sockets++;
   });
+  if(user === undefined) res.redirect("/login");
+  let world = await containerData(user);
+  const proxy = httpProxy.createServer({});
+  proxy.web(
+    req,
+    res,
+    {target: `http://localhost:${world.port}/`}
+  );
   proxy.on("error", (err, req, res) => {
-    console.log(`PROXY ERROR (2): ${err}`);
-    delete registry[user];
+    if (err) throw err;
     res.redirect("/login");
   });
-  socket.on('ping', () => {
+  return;
+});
+
+/**
+ * Handles transfer of HTTP to websocket
+ * @param {Object} req      Web request
+ * @param {Object} socket   Socket created
+ * @param {Object} head     ?
+ */
+app.on("upgrade", async (req, socket, head) => {
+  let user;
+  session(req, {}, () => {
+    user = req.session.user;
+  });
+  let proxy = httpProxy.createServer({});
+  let world = await containerData(user);
+  proxy.ws(
+    req,
+    socket,
+    head,
+    {target: `http://localhost:${world.port}/`}
+  );
+  proxy.on("error", (err, req, res) => {
+    if (err) throw err;
+    res.rediect("/login");
+  });
+  socket.on("ping", () => {
     socket.pong();
   });
-  socket.on('data', () => {
-    emitter.emit('register',
+  socket.on("data", () => {
+    events.emit("lastActive",
       {
-        user: user,
-        params: {
-          active: now()
+        [user]: {
+          lastActive: now()
         }
       }
-    );
+    )
   });
-  socket.on('close', () => {
+  socket.on("close", () => {
     socket.end();
     socket.destroy();
   });
+  return;
 });
 
-/**
- * Event handler for server-side errors
- * @param {String} err  Error message
- */
-server.on("error", err => console.log(err));
+// Activity monitoring
 
-/**
- * Event handler for proxy-side errors
- * @param {String} err  Error message
- */
-app.on("error", err => console.log(err));
+let timeout = process.env.TIMEOUT || 600;
 
-// Remove containers after idle status (TIMEOUT in seconds)
-
-let timeout = process.env.TIMEOUT || 900;
-
-setInterval(() => {
+setInterval( () => {
   const timed = Object
-    .keys(registry)
+    .keys(activity)
     .filter((user, idx, self) => {
-      let lastActive = registry[user].params.active;
-      return now() - lastActive > timeout;
+      return now() - activity[user].lastActive > timeout;
     });
-  for (let entry in timed) {
+  for(let entry in timed) {
     let user = timed[entry];
-    let id = registry[user].params.container.id;
-    emitter.emit('SIGUSER', user, id);
-    delete registry[user];
+    events.emit("SIGUSER", user);
+    delete activity[user];
   }
 }, 10000);
 
-// PRUNE PATROL
+// Prune patrol
 
-setInterval(async () => {
-    let list = await ishmael.listContainers({all: true});
-    let pruned = await ishmael.pruneContainers({until: now()});
-    let banished = pruned['ContainersDeleted'];
-    const remove = Object
-      .keys(registry)
-      .filter((id, idx, self) => {
-        if(banished) {
-          return banished.indexOf(id);
-        }
-      });
-    remove.forEach(elem => {
-        let user = elem;
-        let container = registry[user];
-        emitter.emit('SIGUSER', user, container);
+setInterval( async () => {
+  let list = await moby.listContainers({all: true});
+  let pruned = await moby.pruneContainers({until: now()});
+  let banished = pruned['ContainersDeleted'];
+  const remove = Object
+    .keys(activity)
+    .filter((user, idx, self) => {
+      if(banished) {
+        return banished.indexOf(user);
+      }
     });
+  remove.forEach(user => {
+    emitter.emit("SIGUSER", user);
+  });
 }, 10000);
 
-//Remove the container on SIGINT or exit
+// Container removal
 
 const exit = () => {
   process.exit();
 };
 
-async function spindown(sig) {
-  let user = sig[1];
-  let args = sig[0] == 'USER' ? { filters: {"id":[`${sig[2]}`]} } : {all: true};
-  delete registry[user];
+let interrupt;
+
+const spindown = async (sig) => {
+  let world = await containerData(sig[1]);
+  let args = sig[0] == "USER" ? { filters: {"id":[`${world.id}`]} } : {all: true};
+  delete activity[user];
   if(args.all) { interrupt = true; }
-  let list = await ishmael.listContainers(args);
-  for await (let entry of list) {
-    let container = await ishmael.getContainer(entry.Id);
+  let list = await moby.listContainers(args);
+  for await(let entry of list) {
+    let container = await moby.getContainer(entry.Id);
     let stoppage = await container.stop();
     let removal = await container.remove();
   }
-  let pruned = await ishmael.pruneContainers({until: now()})
+  let pruned = await moby.pruneContainers({until: now()});
   if(args.all) { exit(); }
 }
 
 process.on("SIGINT", spindown.bind());
 process.on("SIGTERM", spindown.bind());
 
-// Nonce custom signal to indicate single user container spindown
+// Nonce signal to indiate single user container shutdown
 
-emitter.on('SIGUSER', (user, id) => {
-  console.log(`SINGLE USER SPIN DOWN: ${user}`)
-  spindown(['USER', user, id]);
+events.on("SIGUSER", (user) => {
+  spindown(["USER", user]);
 });
